+++ aten/src/ATen/native/TensorConversions.cpp
@@ -1707,15 +1707,5 @@ std::vector<Tensor> to_meta(at::ITensorListRef t_list) {
   return outs;
 }
 
-Tensor test_test(const Tensor& x, const Tensor& y) {
-  auto m = x.size(-2);
-  auto n = y.size(-1);
-  auto sizes = at::DimVector(y.sizes().slice(0, y.dim() - 2));
-  sizes.push_back(m);
-  sizes.push_back(n);
-  auto result = at::empty(sizes, y.options());
-  return _compressed_row_strided_mm_out(x, y, result);
-}
-
 } // namespace native
 } // namespace at
